Changed order of execution for sell function in casino.py
To (hopefully) prevent the execution of multiple sell commands, this change would remove gold from the bot's sql db prior to making an API call to GGn. 

Testing is required to verify if this functions as intended.  

If this does not fix the issue, a global thread lock (or a per-user thread lock) could be implemented instead, but that is a more complex solution.